google_doc_api/requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
pybase64
flask